Split the implementation and definitions into individual files
