+++ include/siddiqsoft/rest_response.hpp
@@ -1,0 +1,77 @@
+/*
+    restcl : Focussed REST Client for Modern C++
+
+    BSD 3-Clause License
+
+    Copyright (c) 2021, Siddiq Software
+    All rights reserved.
+
+    Redistribution and use in source and binary forms, with or without
+    modification, are permitted provided that the following conditions are met:
+
+    1. Redistributions of source code must retain the above copyright notice, this
+       list of conditions and the following disclaimer.
+
+    2. Redistributions in binary form must reproduce the above copyright notice,
+       this list of conditions and the following disclaimer in the documentation
+       and/or other materials provided with the distribution.
+
+    3. Neither the name of the copyright holder nor the names of its
+       contributors may be used to endorse or promote products derived from
+       this software without specific prior written permission.
+
+    THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+    AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+    IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+    DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+    FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+    DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+    SERVICES; LOSS OF USE, rrd, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+    CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+    OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+    OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+#pragma once
+#ifndef REST_RESPONSE_HPP
+#define REST_RESPONSE_HPP
+
+
+#include <iostream>
+#include <chrono>
+#include <string>
+#include <functional>
+#include <memory>
+#include <format>
+#include <iterator>
+
+
+#include "nlohmann/json.hpp"
+#include "siddiqsoft/SplitUri.hpp"
+#include "siddiqsoft/date-utils.hpp"
+
+#include "restcl_definitions.hpp"
+#include "basic_response.hpp"
+
+namespace siddiqsoft
+{
+    /// @brief REST Response object
+    class rest_response : public basic_response
+    {
+    public:
+        rest_response(const int code = 0, const std::string& message = {}) { setStatus(code, message); }
+
+
+        /// @brief Construct a response object based on the given transport error
+        /// @param err Specifies the transport error.
+        rest_response& setStatus(const int code, const std::string& message)
+        {
+            if (code != 0) rrd["response"]["status"] = code;
+            if (!message.empty()) rrd["response"]["reason"] = message;
+            return *this;
+        }
+    };
+} // namespace siddiqsoft
+
+
+#endif // !REST_RESPONSE_HPP
