+++ tests/test_unix.cpp
@@ -40,7 +40,6 @@
 
 #include "nlohmann/json.hpp"
 #include "../include/siddiqsoft/restcl.hpp"
-#include "../include/siddiqsoft/restcl_unix.hpp"
 
 
 namespace siddiqsoft
@@ -50,7 +49,7 @@ namespace siddiqsoft
     TEST(TSendRequest, test1a)
     {
         std::atomic_bool passTest = false;
-        HttpRESTClient   wrc;
+        restcl           wrc;
 
         wrc.send("https://www.siddiqsoft.com/"_GET, [&passTest](const auto& req, const auto& resp) {
             nlohmann::json doc(req);
@@ -76,7 +75,7 @@ namespace siddiqsoft
     TEST(TSendRequest, test2a)
     {
         std::atomic_bool passTest = false;
-        HttpRESTClient   wrc(std::format("siddiqsoft.restcl.tests/1.0 (Windows NT; x64; s:{})", __func__));
+        restcl           wrc(std::format("siddiqsoft.restcl.tests/1.0 (Windows NT; x64; s:{})", __func__));
 
         wrc.send("https://reqbin.com/echo/post/json"_OPTIONS, [&passTest](const auto& req, auto& resp) {
             // Checks the implementation of the encode() implementation
@@ -102,8 +101,8 @@ namespace siddiqsoft
         using namespace siddiqsoft::splituri_literals;
         std::atomic_bool passTest = false;
 
-        HttpRESTClient wrc;
-        std::string    responseContentType {};
+        restcl      wrc;
+        std::string responseContentType {};
 
         wrc.send(ReqPost {"https://httpbin.org/post"_Uri,
                           {{"Content-Type", "application/json"}},
@@ -137,7 +136,7 @@ namespace siddiqsoft
         std::atomic_bool passTest = false;
         // auto auth     = base64encode("aau:paau");
 
-        HttpRESTClient wrc(std::format("siddiqsoft.restcl.tests/1.0 (Windows NT; x64; s:{})", __FUNCTION__));
+        restcl wrc(std::format("siddiqsoft.restcl.tests/1.0 (Windows NT; x64; s:{})", __FUNCTION__));
 
         wrc.send(ReqPost {"https://httpbin.org/post"_Uri,
                           {{"Authorization", "Basic YWF1OnBhYXU="}, {"Content-Type", "application/json+custom"}},
@@ -167,7 +166,7 @@ namespace siddiqsoft
         std::atomic_bool passTest = false;
         using namespace siddiqsoft::splituri_literals;
 
-        HttpRESTClient wrc(std::format("siddiqsoft.restcl.tests/1.0 (Windows NT; x64; s:{})", __FUNCTION__));
+        restcl wrc(std::format("siddiqsoft.restcl.tests/1.0 (Windows NT; x64; s:{})", __FUNCTION__));
 
         wrc.send("https://www.siddiqsoft.com:65535/"_GET, [&passTest](const auto& req, const auto& resp) {
             // nlohmann::json doc(req);
@@ -194,7 +193,7 @@ namespace siddiqsoft
         std::atomic_bool passTest = false;
         using namespace siddiqsoft::splituri_literals;
 
-        HttpRESTClient wrc(std::format("siddiqsoft.restcl.tests/1.0 (Windows NT; x64; s:{})", __FUNCTION__));
+        restcl wrc(std::format("siddiqsoft.restcl.tests/1.0 (Windows NT; x64; s:{})", __FUNCTION__));
 
         wrc.send("https://localhost:65535/"_GET, [&passTest](const auto& req, const auto& resp) {
             nlohmann::json doc(req);
@@ -221,7 +220,7 @@ namespace siddiqsoft
         std::atomic_bool passTest = false;
         using namespace siddiqsoft::splituri_literals;
 
-        HttpRESTClient wrc(std::format("siddiqsoft.restcl.tests/1.0 (Windows NT; x64; s:{})", __FUNCTION__));
+        restcl wrc(std::format("siddiqsoft.restcl.tests/1.0 (Windows NT; x64; s:{})", __FUNCTION__));
 
         // The endpoint does not support OPTIONS verb. Moreover, it does not listen on port 9090 either.
         wrc.send("https://httpbin.org:9090/get"_OPTIONS, [&passTest](const auto& req, auto& resp) {
@@ -246,7 +245,7 @@ namespace siddiqsoft
         std::atomic_bool passTest = false;
         using namespace siddiqsoft::splituri_literals;
 
-        HttpRESTClient wrc(std::format("siddiqsoft.restcl.tests/1.0 (Windows NT; x64; s:{})", __FUNCTION__));
+        restcl wrc(std::format("siddiqsoft.restcl.tests/1.0 (Windows NT; x64; s:{})", __FUNCTION__));
 
         wrc.send("https://google.com/"_OPTIONS, [&passTest](const auto& req, auto& resp) {
             // std::cerr << "From callback Wire serialize              : " << req.encode() << std::endl;
@@ -271,7 +270,7 @@ namespace siddiqsoft
     TEST(TSendRequest, test9a)
     {
         std::atomic_bool passTest = false;
-        HttpRESTClient   wrc;
+        restcl           wrc;
 
         wrc.send("https://www.google.com/"_GET, [&passTest](const auto& req, const auto& resp) {
             // std::cerr << "From callback Serialized json: " << req << std::endl;
@@ -293,12 +292,12 @@ namespace siddiqsoft
 
     TEST(restcl, MoveConstructor)
     {
-        std::atomic_uint                        passTest {0};
-        std::vector<siddiqsoft::HttpRESTClient> clients;
+        std::atomic_uint                passTest {0};
+        std::vector<siddiqsoft::restcl> clients;
 
         for (auto i = 0; i < 4; i++) {
             // Move constructor!
-            clients.push_back(siddiqsoft::HttpRESTClient {});
+            clients.push_back(siddiqsoft::restcl {});
         }
 
         EXPECT_EQ(4, clients.size());
@@ -331,7 +330,7 @@ namespace siddiqsoft
         std::cerr << std::format("Starting..\n");
 
         {
-            HttpRESTClient wrc;
+            restcl wrc;
 
             std::cerr << std::format("Post wrc..\n");
             basic_callbacktype valid = [&passTest](const auto& req, const auto& resp) {
@@ -391,8 +390,8 @@ namespace siddiqsoft
             using namespace std::chrono_literals;
             using namespace siddiqsoft::restcl_literals;
 
-            siddiqsoft::HttpRESTClient wrc {"pmd4-drift-check"};
-            nlohmann::json             myStats {"Test", "drift-check"};
+            siddiqsoft::restcl wrc {"pmd4-drift-check"};
+            nlohmann::json     myStats {"Test", "drift-check"};
 
             auto req = "https://time.akamai.com/?iso"_GET;
             if (siddiqsoft::basic_response resp = wrc.send(req); resp.success()) {
